AtlasProperties made unaware of MyDAS
