+++ atlas-utils/src/main/java/uk/ac/ebi/gxa/properties/AtlasProperties.java
@@ -25,16 +25,11 @@ package uk.ac.ebi.gxa.properties;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.util.ReflectionUtils;
 import uk.ac.ebi.gxa.utils.LazyKeylessMap;
-import uk.ac.ebi.mydas.configuration.DataSourceConfiguration;
-import uk.ac.ebi.mydas.configuration.GlobalConfiguration;
-import uk.ac.ebi.mydas.configuration.Mydasserver;
-import uk.ac.ebi.mydas.configuration.ServerConfiguration;
-import uk.ac.ebi.mydas.controller.DataSourceManager;
-import uk.ac.ebi.mydas.controller.MydasServlet;
 
-import java.lang.reflect.Field;
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyVetoException;
+import java.beans.VetoableChangeListener;
 import java.util.*;
 
 /**
@@ -58,6 +53,7 @@ public class AtlasProperties {
     }
 
     private final List<AtlasPropertiesListener> listeners = new ArrayList<AtlasPropertiesListener>();
+    private final List<VetoableChangeListener> vetoableChangeListeners = new ArrayList<VetoableChangeListener>();
     private final Map<String, String> cache = new HashMap<String, String>();
 
     /**
@@ -82,11 +78,13 @@ public class AtlasProperties {
      * @param newValue property value or null if property customization should be deleted
      */
     public void setProperty(String key, String newValue) {
-        if (!key.equals("atlas.dasbase") || updateDasBaseURL(newValue)) {
-            // Update atlas.dasbase property only if we managed to MydasServer code was updated successfully
+        try {
+            notifyListeners(key, getProperty(key), newValue);
             storage.setProperty(key, newValue);
             cache.put(key, newValue);
             notifyListeners();
+        } catch (PropertyVetoException e) {
+            log.warn("Property change vetoed", e);
         }
     }
 
@@ -148,9 +146,35 @@ public class AtlasProperties {
         listeners.remove(listener);
     }
 
+    /**
+     * Register property update listener
+     *
+     * @param listener listener reference
+     */
+    public synchronized void registerListener(VetoableChangeListener listener) {
+        if (!vetoableChangeListeners.contains(listener))
+            vetoableChangeListeners.add(listener);
+    }
+
+    /**
+     * Unregister property update listener
+     *
+     * @param listener listener reference
+     */
+    public synchronized void unregisterListener(VetoableChangeListener listener) {
+        vetoableChangeListeners.remove(listener);
+    }
+
     private synchronized void notifyListeners() {
         for (AtlasPropertiesListener listener : listeners)
             listener.onAtlasPropertiesUpdate(this);
+    }
+
+    private synchronized void notifyListeners(String property, String oldValue, String newValue) throws PropertyVetoException {
+        final PropertyChangeEvent event = new PropertyChangeEvent(this, property, oldValue, newValue);
+        for (VetoableChangeListener listener : vetoableChangeListeners) {
+            listener.vetoableChange(event);
+        }
     }
 
     /* Version properties */
@@ -453,61 +477,5 @@ public class AtlasProperties {
      */
     public Integer getMaxEfoMappingsCountForStructuredQuery() {
         return getIntProperty("atlas.structured.query.max.efo.mappings.count");
-    }
-
-    /**
-     * MydasServlet, used by Atlas to expose its data as a DAS source, is configured at start up via MydasServerConfig.xml.
-     * Maven build replaces atlas.dasbase placeholder in MydasServerConfig.xml with a value set in atlas-web/pom.xml
-     * atlas.dasbase property is also configurable via AtlasProperties, but since MydasServlet code does not currently
-     * provide access to its internal fields using atlas.dasbase, the only current way to re-configure MydasServlet code after
-     * an AtlasProperties change to atlas.dasbase is vai the reflection hack below.
-     * TODO replace this method with direct calls to MydasServlet code once setter methods are provided by the DAS team
-     *
-     * @param dasBaseURL base URL for DAS
-     * @return true if all fields were updated via reflection successfully; false otherwise
-     */
-    public boolean updateDasBaseURL(String dasBaseURL) {
-        boolean success = false;
-        try {
-            Field field = ReflectionUtils.findField(MydasServlet.class, "DATA_SOURCE_MANAGER");
-            field.setAccessible(true);
-            Object dataSourceManager = ReflectionUtils.getField(field, null);
-            if (dataSourceManager != null) {
-                // GxaS4DasDataSource has been accessed at least once since Atlas startup and MydasServerConfig.xml was already
-                // read in by MydasServlet - need to update the relevant object fields via reflection
-                // web.xml is now configured to load MydasServlet at Atlas startup - if it is not loaded by the time
-                // this method runs 
-                ServerConfiguration serverConfiguration = ((DataSourceManager) dataSourceManager).getServerConfiguration();
-
-                // Set baseUrl to dasBaseURL - c.f. <baseurl>${atlas.dasbase}/</baseurl> in MydasServerConfig.xml
-                GlobalConfiguration globalConfiguration = serverConfiguration.getGlobalConfiguration();
-                Field baseUrl = globalConfiguration.getClass().getDeclaredField("baseURL");
-                baseUrl.setAccessible(true);
-                baseUrl.set(globalConfiguration, dasBaseURL);
-                log.debug("Setting <baseurl> MydasServerConfig.xml to: dasBaseURL");
-
-                // Update all capability fields with the new dasBaseURL - c.f. (in MydasServerConfig.xml)
-                //    <capability type="das1:sources" query_uri="${atlas.dasbase}/s4" />
-                //    <capability type="das1:types" query_uri="${atlas.dasbase}/s4/types" />
-                //    <capability type="das1:features" query_uri="${atlas.dasbase}/s4/features?segment=ENSG00000162552" />
-                Map<String, DataSourceConfiguration> dataSourceConfigMap = serverConfiguration.getDataSourceConfigMap();
-                for (DataSourceConfiguration dsConfig : dataSourceConfigMap.values()) {
-                    List<Mydasserver.Datasources.Datasource.Version> versions = dsConfig.getConfig().getVersion();
-                    for (Mydasserver.Datasources.Datasource.Version version : versions) {
-                        List<Mydasserver.Datasources.Datasource.Version.Capability> capabilities = version.getCapability();
-                        for (Mydasserver.Datasources.Datasource.Version.Capability capability : capabilities) {
-                            String queryUri = capability.getQueryUri();
-                            String newQueryUri = queryUri.replaceFirst(".*/", dasBaseURL + "/");
-                            log.debug("Setting query_uri of capability type: " + capability.getType() + " in MydasServerConfig.xml to " + newQueryUri);
-                            capability.setQueryUri(newQueryUri);
-                        }
-                    }
-                }
-                success = true;
-            }
-        } catch (Exception e) {
-            log.error("Failed to update dasBaseUrl to : " + dasBaseURL, e);
-        }
-        return success;
     }
 }
